refactor: economy phase-2 slice 12 — production scenario tables (#3856)
Migrate economy_production_test.dart to table-driven scenarios in
colonizethis_economy_test_support; preserve all eleven test descriptions.

Refs #3856
