+++ packages/colonizethis_economy/REFACTOR_TRACE.md
@@ -308,7 +308,19 @@ Modules:
 
 Module: `colonizethis_economy_test_support/lib/src/build_cost_scenarios.dart`
 
+## Phase 2 — Slice 12 (Refs #3856)
+
+| scenario_id | test description | source file(s) | refs |
+|-------------|------------------|----------------|------|
+| production-recipe-* | five `resolveProduction` recipe/labour rows | `economy_production_test.dart` | — |
+| production-edge-* | three `resolveProduction` edge-case rows | `economy_production_test.dart` | — |
+| effective-labour-* | three `effectiveLabourForWorkers` rows | `economy_production_test.dart` | — |
+
+Module: `colonizethis_economy_test_support/lib/src/economy_production_scenarios.dart`
+
 ## Deferred (phase 2 follow-up slices)
 
-- Remaining core economy suites (`economy_production_test.dart`, `worker_action_cost_test.dart`, `non_gp_auto_offers_test.dart`) → scenario tables
+- Remaining core economy suites (`worker_action_cost_test.dart`, `non_gp_auto_offers_test.dart`) → scenario tables
+- Deal-matcher treasury dedup (`treasury_bid_budget.dart` shared helpers + CI gate)
+- Extended inline-`Game(` guard for all `packages/colonizethis_economy/test/**`
 - ≥15% test LOC reduction target (≤6,500 LOC) — met; maintain on follow-up migrations
